Separate mapProps from script.js
